Combine View classes into edits.py
